Move filtered transaction API/impl to ledger-common-data
