--- components/ledger/ledger-filtered-transaction-api/detekt-baseline.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" ?>
-<SmellBaseline>
-    <ManuallySuppressedIssues></ManuallySuppressedIssues>
-    <CurrentIssues></CurrentIssues>
-</SmellBaseline>
