--- components/ledger/ledger-filtered-transaction-api/src/main/java/net/corda/ledger/filtered/factory/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.ledger.filtered.factory;
-
-import org.osgi.annotation.bundle.Export;
