--- components/ledger/ledger-filtered-transaction-api/src/main/kotlin/net/corda/ledger/filtered/FilteredComponentGroup.kt
@@ -1,13 +1,0 @@
-package net.corda.ledger.filtered
-
-import net.corda.v5.base.annotations.CordaSerializable
-import net.corda.v5.crypto.merkle.MerkleProof
-
-/**
- * A filtered component group.
- *
- * @property componentGroupIndex The index of the component group this [FilteredComponentGroup] represents.
- * @property merkleProof The [MerkleProof] calculated from and containing the filtered components.
- */
-@CordaSerializable
-data class FilteredComponentGroup(val componentGroupIndex: Int, val merkleProof: MerkleProof)
