--- components/ledger/ledger-filtered-transaction/build.gradle
@@ -1,38 +1,0 @@
-plugins {
-    id 'corda.common-publishing'
-    id 'corda.osgi-test-conventions'
-    id 'corda.common-library'
-}
-
-description 'Corda ledger filtered transaction implementation'
-
-configurations {
-    cpis {
-        canBeConsumed = false
-        transitive = false
-    }
-}
-
-ext.cordaEnableFormatting = true
-
-dependencies {
-    compileOnly 'org.osgi:osgi.annotation'
-
-    implementation 'net.corda:corda-ledger-common'
-
-    implementation project(':components:ledger:ledger-filtered-transaction-api')
-
-    implementation project(':libs:serialization:serialization-checkpoint-api')
-    implementation project(':libs:serialization:serialization-internal')
-    implementation project(':libs:crypto:crypto-core')
-    implementation project(':libs:serialization:json-validator')
-    implementation project(':libs:sandbox-types')
-    implementation project(':libs:platform-info')
-
-    runtimeOnly project(":libs:web:web-impl")
-
-    testImplementation project(':libs:serialization:serialization-kryo')
-    testImplementation project(':testing:kryo-serialization-testkit')
-    testImplementation project(':testing:ledger:ledger-common-base-test')
-    testImplementation project(':testing:test-serialization')
-}
