+++ components/ledger/ledger-persistence/build.gradle
@@ -25,7 +25,6 @@ dependencies {
 
     implementation project(':components:configuration:configuration-read-service')
     implementation project(':components:db:db-connection-manager')
-    implementation project(':components:ledger:ledger-filtered-transaction-api')
     implementation project(':components:persistence:persistence-service-common')
     implementation project(':components:virtual-node:cpi-info-read-service')
     implementation project(':components:virtual-node:sandbox-group-context-service')
