+++ components/ledger/ledger-utxo-flow/build.gradle
@@ -28,8 +28,6 @@ dependencies {
     implementation "com.fasterxml.jackson.module:jackson-module-kotlin:$jacksonVersion"
     implementation project(':components:ledger:ledger-common-flow-api')
     implementation project(':components:ledger:ledger-common-flow')
-    implementation project(':components:ledger:ledger-filtered-transaction')
-    implementation project(':components:ledger:ledger-filtered-transaction-api')
     implementation project(':components:ledger:notary-worker-selection')
     implementation project(':components:membership:group-policy')
     implementation project(':components:membership:membership-group-read')
