+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/transaction/filtered/factory/UtxoFilteredTransactionFactoryImpl.kt
@@ -1,7 +1,7 @@
 package net.corda.ledger.utxo.flow.impl.transaction.filtered.factory
 
-import net.corda.ledger.filtered.ComponentGroupFilterParameters
-import net.corda.ledger.filtered.factory.FilteredTransactionFactory
+import net.corda.ledger.common.data.transaction.filtered.ComponentGroupFilterParameters.AuditProof
+import net.corda.ledger.common.data.transaction.filtered.factory.FilteredTransactionFactory
 import net.corda.ledger.utxo.data.transaction.UtxoComponentGroup
 import net.corda.ledger.utxo.data.transaction.UtxoComponentGroup.METADATA
 import net.corda.ledger.utxo.data.transaction.UtxoComponentGroup.NOTARY
@@ -41,10 +41,10 @@ class UtxoFilteredTransactionFactoryImpl @Activate constructor(
         filteredTransactionBuilder: UtxoFilteredTransactionBuilderInternal
     ): UtxoFilteredTransaction {
         val notaryAndTimeWindow = if (filteredTransactionBuilder.notary || filteredTransactionBuilder.timeWindow) {
-            ComponentGroupFilterParameters.AuditProof(
+            AuditProof(
                 NOTARY.ordinal,
                 Any::class.java,
-                ComponentGroupFilterParameters.AuditProof.AuditProofPredicate.Content {
+                AuditProof.AuditProofPredicate.Content {
                     filteredTransactionBuilder.notary && (it is MemberX500Name || it is PublicKey) || // notary components
                         filteredTransactionBuilder.timeWindow && it is TimeWindow // time window
                 }
@@ -57,28 +57,28 @@ class UtxoFilteredTransactionFactoryImpl @Activate constructor(
             filteredTransactionFactory.create(
                 signedTransaction.wireTransaction,
                 listOfNotNull(
-                    ComponentGroupFilterParameters.AuditProof(
+                    AuditProof(
                         METADATA.ordinal,
                         TransactionMetadata::class.java,
-                        ComponentGroupFilterParameters.AuditProof.AuditProofPredicate.Content { true }
+                        AuditProof.AuditProofPredicate.Content { true }
                     ),
                     notaryAndTimeWindow,
                     filteredTransactionBuilder.signatories,
                     filteredTransactionBuilder.inputStates,
                     filteredTransactionBuilder.referenceStates,
-                    (filteredTransactionBuilder.outputStates as? ComponentGroupFilterParameters.AuditProof<*>)?.let { _ ->
-                        ComponentGroupFilterParameters.AuditProof(
+                    (filteredTransactionBuilder.outputStates as? AuditProof<*>)?.let { _ ->
+                        AuditProof(
                             UtxoComponentGroup.OUTPUTS_INFO.ordinal,
                             UtxoOutputInfoComponent::class.java,
-                            ComponentGroupFilterParameters.AuditProof.AuditProofPredicate.Content { true }
+                            AuditProof.AuditProofPredicate.Content { true }
                         )
                     },
                     filteredTransactionBuilder.outputStates,
-                    (filteredTransactionBuilder.commands as? ComponentGroupFilterParameters.AuditProof<*>)?.let { _ ->
-                        ComponentGroupFilterParameters.AuditProof(
+                    (filteredTransactionBuilder.commands as? AuditProof<*>)?.let { _ ->
+                        AuditProof(
                             UtxoComponentGroup.COMMANDS_INFO.ordinal,
                             List::class.java,
-                            ComponentGroupFilterParameters.AuditProof.AuditProofPredicate.Content { true }
+                            AuditProof.AuditProofPredicate.Content { true }
                         )
                     },
                     filteredTransactionBuilder.commands
