+++ processors/flow-processor/build.gradle
@@ -37,7 +37,6 @@ dependencies {
     runtimeOnly project(':components:configuration:configuration-read-service-impl')
     runtimeOnly project(':components:flow:flow-service')
     runtimeOnly project(':components:ledger:ledger-common-flow')
-    runtimeOnly project(':components:ledger:ledger-filtered-transaction')
     runtimeOnly project(':libs:ledger:ledger-common-data')
     runtimeOnly project(':components:ledger:ledger-consensual-flow')
     runtimeOnly project(':components:ledger:ledger-utxo-flow')
